+++ instrumentation/jedis/jedis-4.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jedis/v4_0/JedisRequest.java
@@ -8,12 +8,14 @@ package io.opentelemetry.javaagent.instrumentation.jedis.v4_0;
 import static java.nio.charset.StandardCharsets.UTF_8;
 
 import com.google.auto.value.AutoValue;
+import io.opentelemetry.api.GlobalOpenTelemetry;
+import io.opentelemetry.instrumentation.api.incubator.config.internal.DbConfig;
 import io.opentelemetry.instrumentation.api.incubator.semconv.db.RedisCommandSanitizer;
-import io.opentelemetry.javaagent.bootstrap.internal.AgentCommonConfig;
 import java.net.Socket;
 import java.net.SocketAddress;
 import java.util.ArrayList;
 import java.util.List;
+import javax.annotation.Nullable;
 import redis.clients.jedis.CommandArguments;
 import redis.clients.jedis.Protocol;
 import redis.clients.jedis.args.Rawable;
@@ -23,7 +25,8 @@ import redis.clients.jedis.commands.ProtocolCommand;
 public abstract class JedisRequest {
 
   private static final RedisCommandSanitizer sanitizer =
-      RedisCommandSanitizer.create(AgentCommonConfig.get().isQuerySanitizationEnabled());
+      RedisCommandSanitizer.create(
+          DbConfig.isQuerySanitizationEnabled(GlobalOpenTelemetry.get(), "jedis"));
 
   public static JedisRequest create(ProtocolCommand command, List<byte[]> args) {
     return new AutoValue_JedisRequest(command, args);
@@ -62,14 +65,15 @@ public abstract class JedisRequest {
     return sanitizer.sanitize(getOperationName(), getArgs());
   }
 
-  private SocketAddress remoteSocketAddress;
+  @Nullable private SocketAddress remoteSocketAddress;
 
-  public void setSocket(Socket socket) {
+  public void setSocket(@Nullable Socket socket) {
     if (socket != null) {
       remoteSocketAddress = socket.getRemoteSocketAddress();
     }
   }
 
+  @Nullable
   public SocketAddress getRemoteSocketAddress() {
     return remoteSocketAddress;
   }