+++ instrumentation/jetty/jetty-11.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jetty/v11_0/Jetty11InstrumentationModule.java
@@ -26,6 +26,7 @@ public class Jetty11InstrumentationModule extends InstrumentationModule
 
   @Override
   public ElementMatcher.Junction<ClassLoader> classLoaderMatcher() {
+    // added in Servlet 5.0
     return hasClassesNamed("jakarta.servlet.Servlet");
   }