+++ instrumentation/jetty/jetty-8.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jetty/v8_0/Jetty8HandlerAdvice.java
@@ -32,8 +32,7 @@ public class Jetty8HandlerAdvice {
     }
 
     @Nullable
-    public static AdviceScope start(
-        Object source, HttpServletRequest request, HttpServletResponse response) {
+    public static AdviceScope start(HttpServletRequest request, HttpServletResponse response) {
       Context attachedContext = helper().getServerContext(request);
       if (attachedContext != null) {
         // We are inside nested handler, don't create new span
@@ -63,10 +62,9 @@ public class Jetty8HandlerAdvice {
   @Advice.OnMethodEnter(suppress = Throwable.class)
   @Nullable
   public static AdviceScope onEnter(
-      @Advice.This Object source,
       @Advice.Argument(2) HttpServletRequest request,
       @Advice.Argument(3) HttpServletResponse response) {
-    return AdviceScope.start(source, request, response);
+    return AdviceScope.start(request, response);
   }
 
   @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)