+++ instrumentation/liberty/liberty-20.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/liberty/LibertySingletons.java
@@ -19,7 +19,7 @@ public final class LibertySingletons {
 
   private static final Instrumenter<
           ServletRequestContext<HttpServletRequest>, ServletResponseContext<HttpServletResponse>>
-      INSTRUMENTER =
+      instrumenter =
           AgentServletInstrumenterBuilder.<HttpServletRequest, HttpServletResponse>create()
               .addContextCustomizer(
                   (context, request, attributes) ->
@@ -28,7 +28,7 @@ public final class LibertySingletons {
               .build(INSTRUMENTATION_NAME, Servlet3Accessor.INSTANCE);
 
   private static final LibertyHelper<HttpServletRequest, HttpServletResponse> HELPER =
-      new LibertyHelper<>(INSTRUMENTER, Servlet3Accessor.INSTANCE);
+      new LibertyHelper<>(instrumenter, Servlet3Accessor.INSTANCE);
 
   public static LibertyHelper<HttpServletRequest, HttpServletResponse> helper() {
     return HELPER;