+++ instrumentation/spring/spring-scheduling-3.1/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/spring/scheduling/v3_1/SpringSchedulingTest.java
@@ -186,7 +186,7 @@ class SpringSchedulingTest {
                                         equalTo(EXCEPTION_MESSAGE, "failure"),
                                         satisfies(
                                             EXCEPTION_STACKTRACE,
-                                            value -> value.isInstanceOf(String.class)))),
+                                            val -> val.isInstanceOf(String.class)))),
                 span -> span.hasName("error-handler").hasParent(trace.getSpan(0))));
   }
 }